clipboard-apis/async-write-blobtext-read-blobtext-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard write (Blob/text) -> read (Blob/text) tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
function readBlobAsText(file) {
  return new Promise((resolve) => {
    const blobReader = new FileReader();
    blobReader.addEventListener("loadend", () => {
      resolve(blobReader.result);
    });
    blobReader.readAsText(file);
  });
};

promise_test(async t => {
  const inputMessage = "Clipboard write (Blob/text) -> read (Blob/text) test データ";
  const input = new Blob([inputMessage], {type: 'text/plain'});

  await navigator.clipboard.write(input);
  const output = await navigator.clipboard.read();

  const textOutput = await readBlobAsText(output);

  assert_equals(output.type, "text/plain");
  assert_equals(textOutput, inputMessage);
}, "Verify write and read clipboard (DataTransfer/text)");
</script>
Note: This is a manual test because it writes/reads to the shared system
clipboard and thus cannot be run async with other tests that might interact
with the clipboard.
